test: shard tui gateway s3 clusters
